feat(freeMode): init
