+++ src/angular/src/utils/params-list.ts
@@ -16,13 +16,6 @@ export const paramsList = [
   '_edgeSwipeDetection',
   '_edgeSwipeThreshold',
   '_freeMode',
-  '_freeModeMomentum',
-  '_freeModeMomentumRatio',
-  '_freeModeMomentumBounce',
-  '_freeModeMomentumBounceRatio',
-  '_freeModeMomentumVelocityRatio',
-  '_freeModeSticky',
-  '_freeModeMinimumVelocity',
   '_autoHeight',
   'setWrapperSize',
   'virtualTranslate',
