+++ src/components/core/core-class.js
@@ -331,7 +331,7 @@ class Swiper {
       swiper.updateSlidesClasses();
     }
     let translated;
-    if (swiper.params.freeMode) {
+    if (swiper.params.freeMode.enabled) {
       setTranslate();
       if (swiper.params.autoHeight) {
         swiper.updateAutoHeight();
