+++ src/types/components/public-api.d.ts
@@ -17,3 +17,4 @@ export * from './scrollbar';
 export * from './thumbs';
 export * from './virtual';
 export * from './zoom';
+export * from './free-mode';
