+++ src/vue/params-list.js
@@ -16,13 +16,6 @@ const paramsList = [
   '_edgeSwipeDetection',
   '_edgeSwipeThreshold',
   '_freeMode',
-  '_freeModeMomentum',
-  '_freeModeMomentumRatio',
-  '_freeModeMomentumBounce',
-  '_freeModeMomentumBounceRatio',
-  '_freeModeMomentumVelocityRatio',
-  '_freeModeSticky',
-  '_freeModeMinimumVelocity',
   '_autoHeight',
   'setWrapperSize',
   'virtualTranslate',
