+++ src/vue/swiper.js
@@ -36,13 +36,6 @@ const Swiper = {
     edgeSwipeDetection: { type: [Boolean, String], default: undefined },
     edgeSwipeThreshold: { type: Number, default: undefined },
     freeMode: { type: Boolean, default: undefined },
-    freeModeMomentum: { type: Boolean, default: undefined },
-    freeModeMomentumRatio: { type: Number, default: undefined },
-    freeModeMomentumBounce: { type: Boolean, default: undefined },
-    freeModeMomentumBounceRatio: { type: Number, default: undefined },
-    freeModeMomentumVelocityRatio: { type: Number, default: undefined },
-    freeModeSticky: { type: Boolean, default: undefined },
-    freeModeMinimumVelocity: { type: Number, default: undefined },
     autoHeight: { type: Boolean, default: undefined },
     setWrapperSize: { type: Boolean, default: undefined },
     virtualTranslate: { type: Boolean, default: undefined },
